- silence gcc warning about enum use
git-svn-id: svn://svn.code.sf.net/p/gwyddion/code/trunk/gwyddion@11623 25697fe4-d644-49b8-9d30-be5465f2adf3

+++ modules/file/igorfile.c
@@ -714,7 +714,7 @@ igor_read_data_field(const IgorFile *igorfile,
                                       &power10);
     q = pow10(power10);
 
-    switch (wave5->type) {
+    switch ((guint)wave5->type) {
         case IGOR_INT8:
         {
             const gint8 *ps = (const gint8*)buffer;
